test(radio): migrate legacy
