+++ core/src/components/radio/test/legacy/basic/radio.e2e.ts-snapshots/radio-color-md-ltr-Mobile-Safari-linux.png
