+++ core/src/components/radio/test/legacy/basic/radio.e2e.ts-snapshots/radio-group-checked-ios-ltr-Mobile-Chrome-linux.png
