+++ core/src/components/radio/test/legacy/basic/radio.e2e.ts-snapshots/radio-group-checked-ios-rtl-Mobile-Firefox-linux.png
