+++ core/src/components/radio/test/legacy/basic/radio.e2e.ts-snapshots/radio-group-checked-md-ltr-Mobile-Chrome-linux.png
