+++ core/src/components/radio/test/legacy/basic/radio.e2e.ts-snapshots/radio-group-part-md-ltr-Mobile-Firefox-linux.png
