+++ core/src/components/radio/test/legacy/basic/radio.e2e.ts-snapshots/radio-group-unchecked-ios-ltr-Mobile-Firefox-linux.png
