+++ core/src/components/radio/test/legacy/basic/radio.e2e.ts-snapshots/radio-group-unchecked-md-ltr-Mobile-Chrome-linux.png
